Moved Auth Middleware from Role to Request\Auth
