+++ src/ChurchCRM/Slim/Middleware/Request/Auth/AddEventsRoleAuthMiddleware.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace ChurchCRM\Slim\Middleware\Role;
+namespace ChurchCRM\Slim\Middleware\Request\Auth;
 
 class AddEventsRoleAuthMiddleware extends BaseAuthRoleMiddleware {
