+++ src/ChurchCRM/Slim/Middleware/Request/Auth/AdminRoleAuthMiddleware.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace ChurchCRM\Slim\Middleware\Role;
+namespace ChurchCRM\Slim\Middleware\Request\Auth;
 
 class AdminRoleAuthMiddleware extends BaseAuthRoleMiddleware
 {
