+++ src/ChurchCRM/Slim/Middleware/Request/Auth/EditRecordsRoleAuthMiddleware.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace ChurchCRM\Slim\Middleware\Role;
+namespace ChurchCRM\Slim\Middleware\Request\Auth;
 
 class EditRecordsRoleAuthMiddleware extends BaseAuthRoleMiddleware {
