+++ src/api/routes/calendar/events.php
@@ -5,7 +5,7 @@ use ChurchCRM\Base\EventTypeQuery;
 use ChurchCRM\CalendarQuery;
 use ChurchCRM\Event;
 use ChurchCRM\EventCounts;
-use ChurchCRM\Slim\Middleware\Role\AddEventsRoleAuthMiddleware;
+use ChurchCRM\Slim\Middleware\Request\Auth\AddEventsRoleAuthMiddleware;
 use ChurchCRM\Utils\InputUtils;
 use Slim\Http\Response;
