+++ src/api/routes/people/people-person.php
@@ -5,8 +5,8 @@ use ChurchCRM\dto\Photo;
 use ChurchCRM\ListOptionQuery;
 use ChurchCRM\Slim\Middleware\Request\PersonAPIMiddleware;
 use ChurchCRM\SessionUser;
-use ChurchCRM\Slim\Middleware\Role\DeleteRecordRoleAuthMiddleware;
-use ChurchCRM\Slim\Middleware\Role\EditRecordsRoleAuthMiddleware;
+use ChurchCRM\Slim\Middleware\Request\Auth\DeleteRecordRoleAuthMiddleware;
+use ChurchCRM\Slim\Middleware\Request\Auth\EditRecordsRoleAuthMiddleware;
 use ChurchCRM\Utils\MiscUtils;
 use Slim\Http\Request;
 use Slim\Http\Response;
