+++ src/api/routes/system/system-custom-fields.php
@@ -1,7 +1,7 @@
 <?php
 
 use ChurchCRM\PersonCustomMasterQuery;
-use ChurchCRM\Slim\Middleware\Role\AdminRoleAuthMiddleware;
+use ChurchCRM\Slim\Middleware\Request\Auth\AdminRoleAuthMiddleware;
 use Slim\Http\Request;
 use Slim\Http\Response;
